+++ src/Exception/EnvironmentException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace YaFou\Mapik\Exception;
+
+use Exception;
+
+class EnvironmentException extends Exception
+{
+    
+}